+++ tests/test_core.py
@@ -224,8 +224,6 @@ def test_not_implemented():
         assert not hasattr(pc, name)
     with pytest.raises(NotImplementedError):
         pc.fill_null_forward(dictionary)
-    if pa.__version__ < '13':
-        pytest.skip("requires pyarrow >=13")
     with pytest.raises(NotImplementedError):
         pa.table({'': list('aba')}).group_by([]).aggregate([('', 'first'), ('', 'last')])
     with pytest.raises(ValueError):