+++ tests/test_service.py
@@ -300,6 +300,10 @@ def test_apply(client):
         { columns { zipcode { value(index: -1) } } } }'''
     )
     assert data == {'apply': {'columns': {'zipcode': {'value': 2066562337}}}}
+    data = client.execute(
+        '{ apply(pairwiseDiff: {name: "zipcode"}) { columns { zipcode { value } } } }'
+    )
+    assert data == {'apply': {'columns': {'zipcode': {'value': None}}}}
     data = client.execute('{ apply(rank: {name: "zipcode"}) { row { zipcode } } }')
     assert data == {'apply': {'row': {'zipcode': 1}}}
     data = client.execute(